simple_worker/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
requests==2.28.1
urllib3==1.26.12
celery[redis]
pydub
psycopg2-binary
python-dotenv
google-cloud-storage